Update On Gameplay
General implementation of gameplay is implemented. The inconsistency between # of players and # of answers is checked. The winner is displayed if the user enters "f" when writing the answers of the players. When the correct circumstances are created, the game continues with first, entering answers and then writing correct answers. After these, the differences are displayed and the score table is showed.